requirements.txt
SQLAlchemy==0.9.3
Twisted==13.2.0
cffi==0.8.1
cryptography==0.2.1
futures==2.1.6
parse==1.4.1
pyOpenSSL==0.14
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.6.1
six==1.5.2
stuf==0.9.4
txsockjs==1.2.1
wsgiref==0.1.2
zope.interface==4.1.0
SQLObject==1.5.1
Jinja2==2.7.2